requirements.txt
fastapi
uvicorn
pydantic
python-dotenv

langchain
langchain-community
langchain-google-genai
faiss-cpu

sentence-transformers

pypdf
requests
pytz